+++ sdk/scvmm/azure-resourcemanager-scvmm/CHANGELOG.md
@@ -1,14 +1,22 @@
 # Release History
 
-## 1.1.0-beta.1 (Unreleased)
+## 1.0.0-beta.1 (2024-07-04)
 
-### Features Added
+- Azure Resource Manager Scvmm client library for Java. This package contains Microsoft Azure SDK for Scvmm Management SDK. SCVMM Client. Package tag package-2023-10. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Breaking Changes
 
-### Bugs Fixed
+#### `models.VirtualMachineInstances` was modified
 
-### Other Changes
+* `list(java.lang.String,com.azure.core.util.Context)` was removed
+* `list(java.lang.String)` was removed
+
+### Features Added
+
+#### `models.VirtualMachineInstances` was modified
+
+* `listByArm(java.lang.String)` was added
+* `listByArm(java.lang.String,com.azure.core.util.Context)` was added
 
 ## 1.0.0 (2024-06-26)